🐞 BugFix(#139): 생명력 로직 버그수정
