+++ src/features/quiz/hooks.ts
@@ -1,11 +1,7 @@
 import { PartStatus } from '@/features/learn/types';
-import { useUserHpQuery } from '@/features/user/queries';
-import { isLoggedIn } from '@/features/user/service/authUtils';
-import useUserStore from '@/store/useUserStore';
 import hljs from 'highlight.js';
 import { DependencyList, useState, useLayoutEffect, useEffect } from 'react';
-import toast from 'react-hot-toast';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 
 /**
  * 주어진 코드 문자열을 하이라이트 처리된 HTML로 변환하는 React 커스텀 훅입니다.
@@ -73,29 +69,4 @@ export const useLocationQuizState = () => {
     partId: number;
     partStatus: PartStatus;
   };
-};
-
-type useHpUpdate = (isCorrect: boolean) => void;
-export const useHpUpdate: useHpUpdate = isCorrect => {
-  const { mutate: hpUpdate } = useUserHpQuery.updateHp();
-  const { data: userHp } = useUserHpQuery.getHpWhenLoggedIn();
-  const { user } = useUserStore();
-
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (Number(userHp?.hp) === 0) {
-      toast('목숨이 다 소진되었습니다.');
-      navigate('/');
-    }
-
-    if (isCorrect) return;
-
-    if (isLoggedIn(user) && userHp) {
-      hpUpdate({
-        hp: Number(userHp.hp) - 1,
-        hpStorage: userHp.hpStorage,
-      });
-    }
-  }, [isCorrect]);
 };
